# apps/api/src/openapi/schema.json
{
	"openapi": "3.0.2",
	"info": {
		"title": "EVM Balances API",
		"version": "0.1.0"
	},
	"servers": [
		{
			"url": "http://0.0.0.0:8080/"
		},
		{
			"url": "https://empty-hill-3594.fly.dev/"
		}
	],
	"security": [
		{
			"basic": []
		}
	],
	"components": {
		"securitySchemes": {
			"BasicAuth": {
				"type": "http",
				"scheme": "basic"
			}
		},
		"parameters": {
			"Address": {
				"in": "path",
				"name": "address",
				"required": true,
				"schema": {
					"type": "string",
					"description": "Ethereum address",
					"maxLength": 42,
					"minLength": 42,
					"pattern": "^0x[0-9a-fA-F]{40}$",
					"example": "0x0000000000000000000000000000000000000000"
				}
			},
			"Chain": {
				"in": "path",
				"name": "chain",
				"required": true,
				"schema": {
					"type": "string",
					"description": "Chain name",
					"enum": [
						"ethereum",
						"ropsten",
						"rinkeby",
						"polygon-mainnet",
						"polygon-mumbai",
						"avalanche",
						"optimism-mainnet",
						"arbitrum-mainnet",
						"smartchain"
					],
					"example": "polygon-mainnet"
				}
			},
			"Token": {
				"in": "query",
				"name": "token",
				"required": true,
				"schema": {
					"type": "string",
					"description": "token address",
					"maxLength": 42,
					"minLength": 42,
					"pattern": "^0x[0-9a-fA-F]{40}$",
					"example": "0x0000000000000000000000000000000000000000"
				}
			}
		},
		"schemas": {
			"Balance": {
				"additionalProperties": false,
				"type": "object",
				"properties": {
					"address": {
						"type": "string",
						"description": "Ethereum address",
						"maxLength": 42,
						"minLength": 42,
						"pattern": "^0x[0-9a-fA-F]{40}$"
					},
					"balance": {
						"type": "string",
						"description": "Balance in wei",
						"maxLength": 50,
						"minLength": 0,
						"pattern": "^[0-9]{1,50}$"
					},
					"decimals": {
						"type": "integer",
						"description": "Number of decimals",
						"minimum": 1,
						"maximum": 18
					},
					"symbol": {
						"type": "string",
						"description": "Token symbol",
						"maxLength": 50,
						"minLength": 1,
						"pattern": "^[a-zA-Z0-9]{1,50}$"
					},
					"name": {
						"type": "string",
						"description": "Token name",
						"maxLength": 50,
						"minLength": 1,
						"pattern": "^[a-zA-Z0-9]{1,50}$"
					}
				}
			},
			"Balances": {
				"type": "array",
				"items": {
					"additionalProperties": false,
					"type": "object",
					"properties": {
						"address": {
							"type": "string",
							"description": "Ethereum address",
							"maxLength": 42,
							"minLength": 42,
							"pattern": "^0x[0-9a-fA-F]{40}$"
						},
						"balance": {
							"type": "string",
							"description": "Balance in wei",
							"maxLength": 50,
							"minLength": 0,
							"pattern": "^[0-9]{1,50}$"
						},
						"decimals": {
							"type": "integer",
							"description": "Number of decimals",
							"minimum": 1,
							"maximum": 18
						},
						"symbol": {
							"type": "string",
							"description": "Token symbol",
							"maxLength": 50,
							"minLength": 1,
							"pattern": "^[a-zA-Z0-9]{1,50}$"
						},
						"name": {
							"type": "string",
							"description": "Token name",
							"maxLength": 50,
							"minLength": 1,
							"pattern": "^[a-zA-Z0-9]{1,50}$"
						}
					}
				},
				"maxItems": 1000,
				"minItems": 0
			},
			"Error": {
				"type": "string",
				"description": "Error message",
				"maxLength": 1000,
				"minLength": 0,
				"pattern": "^[a-zA-Z0-9]{1,1000}$"
			},
			"Version": {
				"type": "string",
				"description": "API version",
				"maxLength": 10,
				"minLength": 0,
				"pattern": "^[0-9]{1,10}$"
			}
		}
	},
	"paths": {
		"/version": {
			"get": {
				"responses": {
					"200": {
						"description": "API version",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"version": {
											"type": "string",
											"description": "API version",
											"maxLength": 10,
											"minLength": 0,
											"pattern": "^[0-9]{1,10}$"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "API version",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"version": {
											"type": "string",
											"description": "API version",
											"maxLength": 10,
											"minLength": 0,
											"pattern": "^[0-9]{1,10}$"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/ping": {
			"get": {
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"maxLength": 4,
											"minLength": 4,
											"pattern": "^pong$"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Default response",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"maxLength": 4,
											"minLength": 4,
											"pattern": "^pong$"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/balance/{chain}/{address}": {
			"get": {
				"parameters": [
					{
						"in": "path",
						"name": "chain",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Chain name",
							"enum": [
								"ethereum",
								"ropsten",
								"rinkeby",
								"polygon-mainnet",
								"polygon-mumbai",
								"avalanche",
								"optimism-mainnet",
								"arbitrum-mainnet",
								"smartchain"
							],
							"example": "polygon-mainnet"
						}
					},
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Ethereum address",
							"maxLength": 42,
							"minLength": 42,
							"pattern": "^0x[0-9a-fA-F]{40}$",
							"example": "0x0000000000000000000000000000000000000000"
						}
					},
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string",
							"description": "token address",
							"maxLength": 42,
							"minLength": 42,
							"pattern": "^0x[0-9a-fA-F]{40}$",
							"example": "0x0000000000000000000000000000000000000000"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balance": {
											"additionalProperties": false,
											"type": "object",
											"properties": {
												"address": {
													"type": "string",
													"description": "Ethereum address",
													"maxLength": 42,
													"minLength": 42,
													"pattern": "^0x[0-9a-fA-F]{40}$"
												},
												"balance": {
													"type": "string",
													"description": "Balance in wei",
													"maxLength": 50,
													"minLength": 0,
													"pattern": "^[0-9]{1,50}$"
												},
												"decimals": {
													"type": "integer",
													"description": "Number of decimals",
													"minimum": 1,
													"maximum": 18
												},
												"symbol": {
													"type": "string",
													"description": "Token symbol",
													"maxLength": 50,
													"minLength": 1,
													"pattern": "^[a-zA-Z0-9]{1,50}$"
												},
												"name": {
													"type": "string",
													"description": "Token name",
													"maxLength": 50,
													"minLength": 1,
													"pattern": "^[a-zA-Z0-9]{1,50}$"
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balance": {
											"additionalProperties": false,
											"type": "object",
											"properties": {
												"address": {
													"type": "string",
													"description": "Ethereum address",
													"maxLength": 42,
													"minLength": 42,
													"pattern": "^0x[0-9a-fA-F]{40}$"
												},
												"balance": {
													"type": "string",
													"description": "Balance in wei",
													"maxLength": 50,
													"minLength": 0,
													"pattern": "^[0-9]{1,50}$"
												},
												"decimals": {
													"type": "integer",
													"description": "Number of decimals",
													"minimum": 1,
													"maximum": 18
												},
												"symbol": {
													"type": "string",
													"description": "Token symbol",
													"maxLength": 50,
													"minLength": 1,
													"pattern": "^[a-zA-Z0-9]{1,50}$"
												},
												"name": {
													"type": "string",
													"description": "Token name",
													"maxLength": 50,
													"minLength": 1,
													"pattern": "^[a-zA-Z0-9]{1,50}$"
												}
											}
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Default response",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balance": {
											"additionalProperties": false,
											"type": "object",
											"properties": {
												"address": {
													"type": "string",
													"description": "Ethereum address",
													"maxLength": 42,
													"minLength": 42,
													"pattern": "^0x[0-9a-fA-F]{40}$"
												},
												"balance": {
													"type": "string",
													"description": "Balance in wei",
													"maxLength": 50,
													"minLength": 0,
													"pattern": "^[0-9]{1,50}$"
												},
												"decimals": {
													"type": "integer",
													"description": "Number of decimals",
													"minimum": 1,
													"maximum": 18
												},
												"symbol": {
													"type": "string",
													"description": "Token symbol",
													"maxLength": 50,
													"minLength": 1,
													"pattern": "^[a-zA-Z0-9]{1,50}$"
												},
												"name": {
													"type": "string",
													"description": "Token name",
													"maxLength": 50,
													"minLength": 1,
													"pattern": "^[a-zA-Z0-9]{1,50}$"
												}
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/balances/{chain}/{address}": {
			"get": {
				"parameters": [
					{
						"in": "path",
						"name": "chain",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Chain name",
							"enum": [
								"ethereum",
								"ropsten",
								"rinkeby",
								"polygon-mainnet",
								"polygon-mumbai",
								"avalanche",
								"optimism-mainnet",
								"arbitrum-mainnet",
								"smartchain"
							],
							"example": "polygon-mainnet"
						}
					},
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Ethereum address",
							"maxLength": 42,
							"minLength": 42,
							"pattern": "^0x[0-9a-fA-F]{40}$",
							"example": "0x0000000000000000000000000000000000000000"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balances": {
											"type": "array",
											"items": {
												"additionalProperties": false,
												"type": "object",
												"properties": {
													"address": {
														"type": "string",
														"description": "Ethereum address",
														"maxLength": 42,
														"minLength": 42,
														"pattern": "^0x[0-9a-fA-F]{40}$"
													},
													"balance": {
														"type": "string",
														"description": "Balance in wei",
														"maxLength": 50,
														"minLength": 0,
														"pattern": "^[0-9]{1,50}$"
													},
													"decimals": {
														"type": "integer",
														"description": "Number of decimals",
														"minimum": 1,
														"maximum": 18
													},
													"symbol": {
														"type": "string",
														"description": "Token symbol",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													},
													"name": {
														"type": "string",
														"description": "Token name",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													}
												}
											},
											"maxItems": 1000,
											"minItems": 0
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balances": {
											"type": "array",
											"items": {
												"additionalProperties": false,
												"type": "object",
												"properties": {
													"address": {
														"type": "string",
														"description": "Ethereum address",
														"maxLength": 42,
														"minLength": 42,
														"pattern": "^0x[0-9a-fA-F]{40}$"
													},
													"balance": {
														"type": "string",
														"description": "Balance in wei",
														"maxLength": 50,
														"minLength": 0,
														"pattern": "^[0-9]{1,50}$"
													},
													"decimals": {
														"type": "integer",
														"description": "Number of decimals",
														"minimum": 1,
														"maximum": 18
													},
													"symbol": {
														"type": "string",
														"description": "Token symbol",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													},
													"name": {
														"type": "string",
														"description": "Token name",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													}
												}
											},
											"maxItems": 1000,
											"minItems": 0
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Default",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balances": {
											"type": "array",
											"items": {
												"additionalProperties": false,
												"type": "object",
												"properties": {
													"address": {
														"type": "string",
														"description": "Ethereum address",
														"maxLength": 42,
														"minLength": 42,
														"pattern": "^0x[0-9a-fA-F]{40}$"
													},
													"balance": {
														"type": "string",
														"description": "Balance in wei",
														"maxLength": 50,
														"minLength": 0,
														"pattern": "^[0-9]{1,50}$"
													},
													"decimals": {
														"type": "integer",
														"description": "Number of decimals",
														"minimum": 1,
														"maximum": 18
													},
													"symbol": {
														"type": "string",
														"description": "Token symbol",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													},
													"name": {
														"type": "string",
														"description": "Token name",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													}
												}
											},
											"maxItems": 1000,
											"minItems": 0
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Get balances for address, provide array of token addresses in body",
				"parameters": [
					{
						"in": "path",
						"name": "chain",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Chain name",
							"enum": [
								"ethereum",
								"ropsten",
								"rinkeby",
								"polygon-mainnet",
								"polygon-mumbai",
								"avalanche",
								"optimism-mainnet",
								"arbitrum-mainnet",
								"smartchain"
							],
							"example": "polygon-mainnet"
						}
					},
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Ethereum address",
							"maxLength": 42,
							"minLength": 42,
							"pattern": "^0x[0-9a-fA-F]{40}$",
							"example": "0x0000000000000000000000000000000000000000"
						}
					}
				],
				"requestBody": {
					"description": "Array of token addresses",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"additionalProperties": false,
								"type": "object",
								"properties": {
									"tokens": {
										"type": "array",
										"minItems": 1,
										"maxItems": 1000,
										"items": {
											"type": "string",
											"maxLength": 42,
											"minLength": 42,
											"pattern": "^0x[0-9a-fA-F]{40}$"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balances": {
											"type": "array",
											"items": {
												"additionalProperties": false,
												"type": "object",
												"properties": {
													"address": {
														"type": "string",
														"description": "Ethereum address",
														"maxLength": 42,
														"minLength": 42,
														"pattern": "^0x[0-9a-fA-F]{40}$"
													},
													"balance": {
														"type": "string",
														"description": "Balance in wei",
														"maxLength": 50,
														"minLength": 0,
														"pattern": "^[0-9]{1,50}$"
													},
													"decimals": {
														"type": "integer",
														"description": "Number of decimals",
														"minimum": 1,
														"maximum": 18
													},
													"symbol": {
														"type": "string",
														"description": "Token symbol",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													},
													"name": {
														"type": "string",
														"description": "Token name",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													}
												}
											},
											"maxItems": 1000,
											"minItems": 0
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balances": {
											"type": "array",
											"items": {
												"additionalProperties": false,
												"type": "object",
												"properties": {
													"address": {
														"type": "string",
														"description": "Ethereum address",
														"maxLength": 42,
														"minLength": 42,
														"pattern": "^0x[0-9a-fA-F]{40}$"
													},
													"balance": {
														"type": "string",
														"description": "Balance in wei",
														"maxLength": 50,
														"minLength": 0,
														"pattern": "^[0-9]{1,50}$"
													},
													"decimals": {
														"type": "integer",
														"description": "Number of decimals",
														"minimum": 1,
														"maximum": 18
													},
													"symbol": {
														"type": "string",
														"description": "Token symbol",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													},
													"name": {
														"type": "string",
														"description": "Token name",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													}
												}
											},
											"maxItems": 1000,
											"minItems": 0
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Default",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"description": "Error message",
											"maxLength": 1000,
											"minLength": 0,
											"pattern": "^[a-zA-Z0-9]{1,1000}$"
										},
										"balances": {
											"type": "array",
											"items": {
												"additionalProperties": false,
												"type": "object",
												"properties": {
													"address": {
														"type": "string",
														"description": "Ethereum address",
														"maxLength": 42,
														"minLength": 42,
														"pattern": "^0x[0-9a-fA-F]{40}$"
													},
													"balance": {
														"type": "string",
														"description": "Balance in wei",
														"maxLength": 50,
														"minLength": 0,
														"pattern": "^[0-9]{1,50}$"
													},
													"decimals": {
														"type": "integer",
														"description": "Number of decimals",
														"minimum": 1,
														"maximum": 18
													},
													"symbol": {
														"type": "string",
														"description": "Token symbol",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													},
													"name": {
														"type": "string",
														"description": "Token name",
														"maxLength": 50,
														"minLength": 1,
														"pattern": "^[a-zA-Z0-9]{1,50}$"
													}
												}
											},
											"maxItems": 1000,
											"minItems": 0
										}
									}
								}
							}
						}
					}
				}
			}
		}
	}
}